+++ src/backend/worksheet/Worksheet.h
@@ -54,8 +54,8 @@ public:
 	enum class Layout {NoLayout, VerticalLayout, HorizontalLayout, GridLayout};
 	enum class CartesianPlotActionMode {ApplyActionToSelection, ApplyActionToAll};
 
-	static float convertToSceneUnits(const float value, const Worksheet::Unit unit);
-	static float convertFromSceneUnits(const float value, const Worksheet::Unit unit);
+	static double convertToSceneUnits(const double value, const Worksheet::Unit unit);
+	static double convertFromSceneUnits(const double value, const Worksheet::Unit unit);
 
 	QIcon icon() const override;
 	QMenu* createContextMenu() override;
